add profile app

+++ django_project/posts/views.py
@@ -6,7 +6,7 @@ from django.shortcuts import get_object_or_404
 
 from . import serializers
 from . import permissions
-from .models import Posts
+from .models import Posts, LikePosts, Comment
 from config.response import set_receive 
 # Create your views here.
 
@@ -42,9 +42,15 @@ class DetailPost(APIView):
         return Response(data)
 
 class ActionPost(APIView):
+    def get_object(self, pk):
+        return get_object_or_404(Posts, pk=pk)
+            
     def put(self, request, pk):
         receive = set_receive(request, request.content_type)
         obj = self.get_object(pk)
+
+        if self.request.user != obj.user:
+            return Response('You not permission!!!', status=403)
 
         obj.content = request.FILES.get('content') if request.FILES.get('content') else obj.content
         obj.caption = receive.get('caption') if receive.get('caption') else obj.caption
@@ -54,6 +60,10 @@ class ActionPost(APIView):
     
     def delete(self, request, pk):
         obj = self.get_object(pk)
+
+        if self.request.user != obj.user:
+            return Response('You not permission!!!', status=403)
+        
         obj.delete()
         return Response('The post was deleted.')
 
@@ -92,3 +102,43 @@ class ListSavedPosts(APIView):
 
         data = serializers.ListPostsSerializer(posts, many=True).data if posts else None
         return Response(data)
+    
+
+
+# ---- Explore ---- 
+class Explore(APIView):
+    pass
+
+
+class LikePost(APIView):
+    def post(self, request):
+        receive = set_receive(request, request.content_type)
+        if receive.get('post_id') is None:
+            return Response('The post_id field is required.')
+        
+        post_obj = get_object_or_404(Posts, id=receive.get('post_id'))
+        if LikePosts.objects.filter(user=request.user, post=post_obj):
+            LikePosts.objects.get(user=request.user, post=post_obj).delete()
+
+        else:
+            LikePosts.objects.create(user=request.user, post=post_obj)
+
+        return Response(status=200)
+    
+class CommentPost(APIView):
+    def post(self, request):
+        receive = set_receive(request, request.content_type)
+
+        if receive.get('post_id') is None or receive.get('text') is None:
+            return Response('The post_id and comment field is required.')
+        
+        if receive.get('reply_comment') is None:
+            create_comment = Comment.objects.create(user=request.user, 
+                post=get_object_or_404(Posts, id=receive.get('post_id')), text=receive.get('text'))
+        
+        else:
+            create_comment = Comment.objects.create(user=request.user, 
+                post=get_object_or_404(Posts, id=receive.get('post_id')), text=receive.get('text'), 
+                reply_comment=get_object_or_404(Comment, id=receive.get('reply_comment')))
+        
+        return Response('The comment was created successfully.')
